refactor(kmb): extract ETA grouping logic to shared module
- Create lib/eta/kmb-eta-groups.ts with groupEtasByVariant() and precomputeRenderGroups()
- Export EtaGroup and PrecomputedGroups types from shared module
- Add buildKey parameter for custom grouping (used by results-kmb.tsx for multipleStops)
- Remove duplicated grouping logic from kmb-pane.tsx and results-kmb.tsx

Closes #70
